auth error update

+++ MySavings/app/src/main/java/com/fredy/mysavings/Feature/Domain/Repository/SyncRepository.kt
@@ -17,7 +17,6 @@ import com.fredy.mysavings.Util.DefaultData.transferCategory
 import com.fredy.mysavings.Util.Log
 import com.fredy.mysavings.Util.isInternetConnected
 import com.google.firebase.auth.FirebaseAuth
-import com.google.firebase.firestore.FirebaseFirestore
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.flow.first
 import kotlinx.coroutines.flow.firstOrNull
@@ -46,16 +45,20 @@ class SyncRepositoryImpl @Inject constructor(
 ) : SyncRepository {
     override suspend fun syncBooks(withDelete: Boolean) {
         withContext(Dispatchers.IO) {
+            Log.i("syncBooks: Start")
             val currentUser = firebaseAuth.currentUser
             currentUser?.let {
                 val userId = if (currentUser.isNotNull()) currentUser.uid else ""
                 try {
-                    val books = bookDataSource.getUserBooksOrderedByName(userId).first()
+                    val books = bookDataSource.getUserBooksOrderedByName(userId).firstOrNull()
+                    if (books.isNullOrEmpty()){
+                        throw(Exception("book is null or empty"))
+                    }
                     if (withDelete) {
                         bookDao.deleteAllBooks()
                     }
                     bookDao.upsertAllBookItem(books)
-                }catch (e:Exception){
+                } catch (e: Exception) {
                     val tempBook = defaultBook.copy(
                         bookId = userId,
                         userIdFk = userId
@@ -65,11 +68,13 @@ class SyncRepositoryImpl @Inject constructor(
                     )
                 }
             }
+            Log.i("syncBooks: Finish")
         }
     }
 
     override suspend fun syncAccounts(withDelete: Boolean) {
         withContext(Dispatchers.IO) {
+            Log.i("syncAccounts: Start")
             val currentUser = firebaseAuth.currentUser
             currentUser?.let {
                 val userId = if (currentUser.isNotNull()) currentUser.uid else ""
@@ -85,11 +90,13 @@ class SyncRepositoryImpl @Inject constructor(
                 }
                 accountDao.upsertAllAccountItem(accounts)
             }
+            Log.i("syncAccounts: Finish")
         }
     }
 
     override suspend fun syncRecords(withDelete: Boolean) {
         withContext(Dispatchers.IO) {
+            Log.i("syncRecords: Start")
             val currentUser = firebaseAuth.currentUser
             currentUser?.let {
                 val userId = if (currentUser.isNotNull()) currentUser.uid else ""
@@ -99,11 +106,13 @@ class SyncRepositoryImpl @Inject constructor(
                 }
                 recordDao.upsertAllRecordItem(records)
             }
+            Log.i("syncRecords: Finish")
         }
     }
 
     override suspend fun syncCategory(withDelete: Boolean) {
         withContext(Dispatchers.IO) {
+            Log.i("syncCategory: Start")
             val currentUser = firebaseAuth.currentUser
             currentUser?.let {
                 val userId = if (currentUser.isNotNull()) currentUser.uid else ""
@@ -125,6 +134,7 @@ class SyncRepositoryImpl @Inject constructor(
                 }
                 categoryDao.upsertAllCategoryItem(categories)
             }
+            Log.i("syncCategory: Finish")
         }
     }
